+++ test/parse.test.js
@@ -166,3 +166,23 @@ describe('bigNarrowPeak', () => {
     expect(f1).toMatchSnapshot()
   })
 })
+
+test('real world', () => {
+  const autoSql = `table hg18KGchr7
+    "UCSC Genes for chr7 with color plus GeneSymbol and SwissProtID"
+    (
+    string  chrom;              "Reference sequence chromosome or scaffold"
+    uint    chromStart; "Start position of feature on chromosome"
+    uint    chromEnd;   "End position of feature on chromosome"
+    string  name;               "Name of gene"
+    uint    score;              "Score"
+    char[1] strand;             "+ or - for strand"
+    uint    thickStart; "Coding region start"
+    uint    thickEnd;   "Coding region end"
+    uint        reserved;       "Green on + strand, Red on - strand"
+    string  geneSymbol; "Gene Symbol"
+    string  spID;               "SWISS-PROT protein Accession number"
+    )`
+  const p = new BED({ autoSql })
+  expect(p.format).toMatchSnapshot()
+})